+++ admin-vben/src/components/ui/card/CardHeader.vue
@@ -1,0 +1,18 @@
+<template>
+  <div :class="cn('flex flex-col space-y-1.5 p-6', props.class)" v-bind="$attrs">
+    <slot />
+  </div>
+</template>
+
+<script setup lang="ts">
+import { cn } from '@/lib/utils';
+
+const props = withDefaults(
+  defineProps<{
+    class?: string;
+  }>(),
+  {
+    class: '',
+  }
+);
+</script>